User fastapi directly instead of uvicorn
