+++ Dockerfile
@@ -30,4 +30,4 @@ USER appuser
 EXPOSE 8000
 
 # Run the command to start Uvicorn
-CMD ["uvicorn", "src.main:app", "--host", "0.0.0.0", "--port", "8000"]
+CMD ["fastapi", "run", "src/main.py", "--host", "0.0.0.0", "--port", "8000"]
