+++ pyproject.toml
@@ -21,7 +21,6 @@ package-mode = false
 python = "^3.12"
 # Add your production dependencies here
 fastapi = "^0.111.0"
-uvicorn = { extras = ["standard"], version = "^0.30.1" }
 asgi-correlation-id = "^4.3.1"
 pytest = "^8.2.2"
 langchain-community = "^0.2.6"
