+++ src/main.py
@@ -1,13 +1,18 @@
-import uvicorn
+from fastapi import FastAPI
+from starlette.responses import StreamingResponse
 
-from src.routes.api import app  # noqa: F401 - uvicorn needs to import the app
+from services.chat import init_chat, process_chat_request
 
-if __name__ == "__main__":
-    config = uvicorn.Config(
-        app="main:app",
-        port=8000,
-        log_level="info",
-        use_colors=True,
-    )
-    server = uvicorn.Server(config)
-    server.run()
+app = FastAPI()
+
+
+@app.get("/chat/init")
+async def init() -> dict:
+    """ Endpoint to initialize the chat with the Kyma companion """
+    return await init_chat()
+
+
+@app.get("/chat")
+async def chat() -> StreamingResponse:
+    """ Endpoint to chat with the Kyma companion """
+    return StreamingResponse(process_chat_request(), media_type='text/event-stream')
